requirements.txt
flask>=3.0
pytest>=8.0
scrapfly-sdk>=0.8
